% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlbum.R
\name{getAlbum}
\alias{getAlbum}
\title{Get additional album info}
\usage{
getAlbum(album_id, market = "US", token)
}
\arguments{
\item{album_id}{The Spotify ID for the artist.}

\item{market}{Optional. An ISO 3166-1 alpha-2 country code. Supply this parameter to limit the response to one particular geographical market. For example, for albums available in Sweden: market=SE.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\value{
'Get an Artist’s Albums
}
\description{
Get additional album info

function to get additional album info from an artist using their ID
}
