% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArtist.R
\name{getArtist}
\alias{getArtist}
\title{Get Spotify catalog information for a single artist}
\usage{
getArtist(artist_id, token)
}
\arguments{
\item{artist_id}{The Spotify ID for the artist.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\value{
Get Spotify catalog information for a single track identified by its unique Spotify ID.
}
\description{
Get Spotify catalog information for a single artist

Function to get Spotify catalog information for a single artist identified by their unique Spotify ID.
}
