% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArtistTopTracks.R
\name{getArtistTopTracks}
\alias{getArtistTopTracks}
\title{Get top tracks of an Artist}
\usage{
getArtistTopTracks(artist_id, country = "US", token)
}
\arguments{
\item{artist_id}{Artist ID}

\item{country}{Required. The country: an ISO 3166-1 alpha-2 country code.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\description{
Get top tracks of an Artist


function to get top tracks of an Artist by artist ID and country
}
