% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAudioFeatures.R
\name{getAudioFeatures}
\alias{getAudioFeatures}
\title{Get audio feature information for a single track identified by its unique Spotify ID.}
\usage{
getAudioFeatures(track_id, token)
}
\arguments{
\item{track_id}{The Spotify ID for the track.}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\description{
Get audio feature information for a single track identified by its unique Spotify ID.


function to get audio feature information for a single track identified by its unique Spotify ID.
}
