% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlaylistsTracks.R
\name{getPlaylistsTracks}
\alias{getPlaylistsTracks}
\title{Get the songs of an specific playlist}
\usage{
getPlaylistsTracks(user_id, playlist_id, offset = 0, limit = 100, token)
}
\arguments{
\item{user_id}{Owner of the playlist ID (username)}

\item{playlist_id}{Playlist ID}

\item{offset}{The index of the first songs to return. Default: 0 (the first object). Maximum offset: 100.000.}

\item{limit}{The limit of result that brings each request}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\description{
Get the songs of an specific playlist


function to get songs about a specifc playlist
}
