% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelatedArtists.R
\name{getRelatedArtist}
\alias{getRelatedArtist}
\title{Get Related artists}
\usage{
getRelatedArtist(artist_id, token)
}
\arguments{
\item{artist_id}{Id of the artist}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\description{
Get Related artists


function to get the related artists of an Artist
}
