% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUser.R
\name{getUser}
\alias{getUser}
\title{Get basic info of an User}
\usage{
getUser(user_id, token)
}
\arguments{
\item{user_id}{user id}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\description{
Get basic info of an User


function to get basic info of an user
}
