% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserPlaylists.R
\name{getUserPlaylists}
\alias{getUserPlaylists}
\title{Get a list of the playlists owned or followed by a Spotify user}
\usage{
getUserPlaylists(username, limit = 50, token)
}
\arguments{
\item{username}{The username}

\item{limit}{The number of results per request. Max 50}

\item{token}{An OAuth token created with \code{spotifyOAuth}.}
}
\value{
get playlist from a specific user using a username
}
\description{
Get a list of the playlists owned or followed by a Spotify user


function to get Info about the playlist of a specific user using a username
}
